+++ test/rendercli/RaytracerOptionTest.cmake
@@ -199,6 +199,19 @@ rendercli_assert_image_dimensions("${wavefront_pathtracer_output}" 24 24
 rendercli_assert_image_nonempty("${wavefront_pathtracer_output}"
                                 NAME "wavefront pathtracer direct pixels")
 
+set(wavefront_denoise_output "${TEST_OUTPUT_DIR}/wavefront-denoise-direct.png")
+rendercli_run(
+  NAME "rendercli wavefront direct engine accepts denoiser"
+  COMMAND
+    "${RENDERCLI}" --direct_engine --engine wavefront --wavefront_denoiser box
+    --wavefront_denoise_radius 1 --width 24 --height 24 --sampler Regular
+    --samples_per_pixel 4 "${raytracer_scene}" "${wavefront_denoise_output}"
+)
+rendercli_assert_image_dimensions("${wavefront_denoise_output}" 24 24
+                                  NAME "wavefront denoise direct dimensions")
+rendercli_assert_image_nonempty("${wavefront_denoise_output}"
+                                NAME "wavefront denoise direct pixels")
+
 set(samples_output "${TEST_OUTPUT_DIR}/samples-per-pixel.png")
 rendercli_run(
   NAME "rendercli raytracer accepts --samples_per_pixel"